+++ R/dummy_functions.R
@@ -23,14 +23,9 @@ example_cdisc_data <- function() { # nolint
   ADSL$logical_test <- sample(c(TRUE, FALSE, NA), size = nrow(ADSL), replace = TRUE) # nolint
   ADSL$SEX[c(2, 5)] <- NA # nolint
 
-  cdisc_data_obj <- teal.data::cdisc_data(
-    cdisc_dataset(dataname = "ADSL", x = ADSL),
-    cdisc_dataset(dataname = "ADTTE", x = ADTTE)
-  )
-
   res <- teal.data::cdisc_data(
-    teal.data::cdisc_dataset(dataname = "ADSL", x = ADSL),
-    teal.data::cdisc_dataset(dataname = "ADTTE", x = ADTTE),
+    ADSL = ADSL,
+    ADTTE = ADTTE,
     code = '
       ADSL <- data.frame(
         STUDYID = "study",
@@ -62,7 +57,7 @@ example_cdisc_data <- function() { # nolint
 #' @keywords internal
 example_datasets <- function() { # nolint
   dummy_cdisc_data <- example_cdisc_data()
-  datasets <- teal.slice::init_filtered_data(dummy_cdisc_data)
+  datasets <- teal_data_to_filtered_data(dummy_cdisc_data)
   list(
     "d2" = list(
       "d3" = list(
@@ -101,13 +96,21 @@ example_module <- function(label = "example teal module", datanames = "all") {
       checkmate::assert_class(data, "tdata")
       moduleServer(id, function(input, output, session) {
         output$text <- renderPrint(data[[input$dataname]]())
+        teal.widgets::verbatim_popup_srv(
+          id = "rcode",
+          verbatim_content = attr(data, "code")(),
+          title = "Association Plot"
+        )
       })
     },
     ui = function(id, data) {
       ns <- NS(id)
       teal.widgets::standard_layout(
         output = verbatimTextOutput(ns("text")),
-        encoding = selectInput(ns("dataname"), "Choose a dataset", choices = names(data))
+        encoding = div(
+          selectInput(ns("dataname"), "Choose a dataset", choices = names(data)),
+          teal.widgets::verbatim_popup_ui(ns("rcode"), "Show R code")
+        )
       )
     },
     datanames = datanames